EVBX-01: add unit tests

+++ src/main/java/com/evbx/resource/layer/service/impl/SpecServiceImpl.java
@@ -66,6 +66,9 @@ public class SpecServiceImpl implements SpecService {
 
     @Override
     public void deleteById(long id) {
+        if (!specRepository.existsById(id)) {
+            throw new ItemNotFoundException(Item.SPECIFICATION, id);
+        }
         specRepository.deleteById(id);
     }
